:truck: Moved global search to fetch
