+++ aksel.nav.no/website/app/_ui/global-search/GlobalSearch.types.ts
